Split Components.kt into individual files and add @Preview to MainScreen and TouchScreen
