+++ sample/src/main/java/com/github/nisrulz/senseysample/ui/Colors.kt
@@ -1,0 +1,8 @@
+package com.github.nisrulz.senseysample.ui
+
+import androidx.compose.ui.graphics.Color
+
+internal val PrimaryBlue = Color(0xFF35518C)
+internal val AccentOrange = Color(0xFFFF9800)
+internal val White = Color(0xFFFFFFFF)
+internal val DividerGray = Color(0xFFB6B6B6)
