+++ sample/src/main/java/com/github/nisrulz/senseysample/ui/MainScreen.kt
@@ -18,6 +18,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 
@@ -26,6 +27,20 @@ data class SensorItem(
     val isSelected: Boolean,
     val onSelect: () -> Unit,
 )
+
+@Preview
+@Composable
+private fun MainScreenPreview() {
+    MainScreen(
+        sensors =
+            listOf(
+                SensorItem("Shake Detector", true, {}),
+                SensorItem("Flip Detector", false, {}),
+            ),
+        resultText = "No gesture detected",
+        onTouchDetectorClick = {},
+    )
+}
 
 @Composable
 fun MainScreen(
