+++ sample/src/main/java/com/github/nisrulz/senseysample/ui/TouchScreen.kt
@@ -11,7 +11,20 @@ import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.tooling.preview.Preview
 import com.github.nisrulz.sensey.gesture.compose.senseyGestures
+
+@Preview
+@Composable
+private fun TouchScreenPreview() {
+    TouchScreen(
+        touchDetectionChecked = true,
+        pinchScaleChecked = false,
+        onTouchDetectionToggle = {},
+        onPinchScaleToggle = {},
+        resultText = "Touch gesture detected",
+    )
+}
 
 @Composable
 fun TouchScreen(
